+++ database/migrations/2020_02_05_123250_create_mensagems_table.php
@@ -21,7 +21,7 @@ class CreateMensagemsTable extends Migration
             $table->date('data');
             $table->string('categoriaRemetente');
 
-            $table->integer('comissaoId');
+            $table->unsignedBigInteger('comissaoId');
         });
     }
