+++ database/migrations/2020_02_05_123453_create_coautors_table.php
@@ -22,7 +22,7 @@ class CreateCoautorsTable extends Migration
             $table->foreign('eventos_id')->references('id')->on('eventos');
 
             $table->unsignedBigInteger('autorId');
-            $table->integer('trabalhoId')->nullable();
+            $table->unsignedBigInteger('trabalhoId')->nullable();
 
         });
     }
