+++ database/migrations/2020_02_05_123532_create_pertences_table.php
@@ -17,8 +17,8 @@ class CreatePertencesTable extends Migration
             $table->bigIncrements('id');
             $table->timestamps();
 
-            $table->integer('revisorId');
-            $table->integer('areaId');
+            $table->unsignedBigInteger('revisorId');
+            $table->unsignedBigInteger('areaId');
         });
     }
