+++ database/migrations/2022_03_26_181753_add_tipo_comissao_to_certificados.php
@@ -14,7 +14,7 @@ class AddTipoComissaoToCertificados extends Migration
     public function up()
     {
         Schema::table('certificados', function (Blueprint $table) {
-            $table->integer('tipo_comissao_id')->nullable();
+            $table->unsignedBigInteger('tipo_comissao_id')->nullable();
         });
     }
