+++ database/migrations/2022_05_31_142504_revisor_id_nullable_in_parecers.php
@@ -14,7 +14,7 @@ class RevisorIdNullableInParecers extends Migration
     public function up()
     {
         Schema::table('parecers', function (Blueprint $table) {
-            $table->integer('revisorId')->nullable()->change();
+            $table->unsignedBigInteger('revisorId')->nullable()->change();
             $table->boolean('parecer_final')->default(false);
         });
     }
